# docs/swagger.yaml
definitions:
...
  friend.UserInfo:
    properties:
      applyTime:
        type: string
      birth:
        type: string
      email:
        type: string
      ex:
        type: string
      flag:
        type: integer
      gender:
        type: integer
      icon:
        type: string
      mobile:
        type: string
      name:
        type: string
      reqMessage:
        type: string
      uid:
        type: string
    type: object
  friend.blackListUserInfo:
    properties:
      birth:
        type: string
      email:
        type: string
      ex:
        type: string
      gender:
        type: integer
      icon:
        type: string
      mobile:
        type: string
      name:
        type: string
      uid:
        type: string
    type: object
...
  friend.paramsAddFriend:
    properties:
      operationID:
        type: string
      reqMessage:
        type: string
      uid:
        type: string
    required:
    - operationID
    - uid
    type: object
...
  friend.paramsGetApplyList:
    properties:
      operationID:
        type: string
    required:
    - operationID
    type: object
...
  friend.paramsImportFriendReq:
    properties:
      operationID:
        type: string
      ownerUid:
        type: string
      uidList:
        items:
          type: string
        type: array
    required:
    - operationID
    - ownerUid
    - uidList
    type: object
  friend.paramsSearchFriend:
    properties:
      operationID:
        type: string
      ownerUid:
        type: string
      uid:
        type: string
    required:
    - operationID
    - uid
    type: object
...
  friend.resultImportFriend:
    properties:
      errCode:
        example: 0
        type: integer
      errMsg:
        example: error
        type: string
      failedUidList:
        items:
          type: string
        type: array
    type: object
...
  user.result:
    properties:
      data: {}
      errCode:
        example: 0
        type: integer
      errMsg:
        example: error
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /friend/add_friend:
    post:
      consumes:
      - application/json
      description: add a new friend by uid
      parameters:
      - description: add friend params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsAddFriend'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
...
  /friend/delete_friend:
    post:
      consumes:
      - application/json
      description: delete friend
      parameters:
      - description: delete friend params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsSearchFriend'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
  /friend/get_blacklist:
    post:
      consumes:
      - application/json
      description: get black list
      parameters:
      - description: get black list
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsSearchFriend'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/friend.blackListUserInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
  /friend/get_friend_apply_list:
    post:
      consumes:
      - application/json
      description: get friend apply list
      parameters:
      - description: get friend apply list
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsGetApplyList'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/friend.UserInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
...
  /friend/get_self_apply_list:
    post:
      consumes:
      - application/json
      description: get friend apply list
      parameters:
      - description: get friend apply list
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsGetApplyList'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/user.result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/friend.UserInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
  /friend/import_friend:
    post:
      consumes:
      - application/json
      description: import friend
      parameters:
      - description: import friend params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/friend.paramsImportFriendReq'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/friend.resultImportFriend'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.result'
      tags:
      - friend
...
swagger: "2.0"